+++ src/constants/index.js
@@ -1,6 +1,8 @@
 export const USERNAME_PREFIX = '--username';
 export const DEFAULT_USERNAME = 'Anonymous';
 
+export const UP_INDICATOR = '..';
+
 export const BROTLI_EXTNAME = '.br';
 
 export const GHZ = 'GHz';
@@ -13,14 +15,12 @@ export const DIRENT_PROPS = {
     NAME: 'name',
     TYPE: 'type',
 };
-
 export const DIRENT_TYPES = {
     FILE: 'file',
     DIRECTORY: 'directory',
 };
 
 export const EXIT_COMMAND = '.exit';
-
 export const OS_COMMAND_ARGUMENTS = {
     EOL: '--EOL',
     CPUS: '--cpus',
