+++ src/modules/commands/ts/cd/index.js
@@ -1,8 +1,8 @@
 import { resolve } from 'path';
 import { readdir } from 'fs/promises';
 
-import { appData } from '../../appData.js';
-import { errorHandler } from '../../errorHandler.js';
+import { appData } from '../../../appData.js';
+import { errorHandler } from '../../../errorHandler.js';
 
 export const cd = async (pathToDir) => {
     const cwd = resolve(appData.cwd, pathToDir);
