+++ src/modules/commands/ts/index.js
@@ -1,0 +1,3 @@
+export { cd } from './cd/index.js';
+export { up } from './up/index.js';
+export { ls } from './ls/index.js';
