+++ src/modules/commands/ts/ls/index.js
@@ -1,11 +1,11 @@
 import { readdir } from 'fs/promises';
 
-import { appData } from '../../appData.js';
-import { errorHandler } from '../../errorHandler.js';
+import { appData } from '../../../appData.js';
+import { errorHandler } from '../../../errorHandler.js';
 
 import { sortDirents } from './helpers/sortDirents.js';
 
-import { DIRENT_TYPES } from '../../../constants/index.js';
+import { DIRENT_TYPES } from '../../../../constants/index.js';
 
 export const ls = async () => {
     try {
