+++ src/modules/commands/ts/up/index.js
@@ -1,0 +1,10 @@
+import { resolve } from 'path';
+
+import { appData } from '../../../appData.js';
+
+import { UP_INDICATOR } from '../../../../constants/index.js';
+
+export const up = () => {
+    appData.cwd = resolve(appData.cwd, UP_INDICATOR);
+};
+
