--- src/modules/commands/up/index.js
@@ -1,8 +1,0 @@
-import { resolve } from 'path';
-
-import { appData } from '../../appData.js';
-
-export const up = () => {
-    appData.cwd = resolve(appData.cwd, '..');
-};
-
